Calculate pseudobulk values

This function will generate a set of pseudobulk expression matrices.

getPseudobulk(
  object,
  replicate_labels,
  split_labels = NULL,
  use_cells = NULL,
  min_cells_per_split = 1,
  min_replicates_per_split = 1,
  min_cells_per_feature = 0,
  use_assay = NULL,
  use_layer = NULL,
  n_cores = NULL,
  verbose = TRUE
)

Arguments

object

An object of class 'Seurat' or 'SingleCellExperiment'.

replicate_labels

A character string or vector indicating the name of the column containing the replicate labels.

split_labels

A character string or vector indicating the name of a column by which to split the cells prior to pseudobulking and performing differential expression (e.g., cell types). Results will be returned for each unique value in the column indicated by 'split_labels'. Default = NULL will run pseudobulk differential expression on all cells together.

use_cells

A vector of cell names subset to. Default = NULL will use all cells.

min_cells_per_split

A numeric value indicating the minimum number of cells within one split. Pseudobulk expression will not be calculated for splits with fewer cells. Defaults to 1.

min_replicates_per_split

A numeric value indicating the minimum number of distinct replicates represented within one split. Pseudobulk expression will not be calculated for splits with fewer replicates. Defaults to 1.

min_cells_per_feature

A numeric value indicating the minimum number of cells (within a split) with expression of a gene. Pseudobulk expression will not be calculated for genes with fewer cells. Defaults to 0.

use_assay

A character string indicating the assay to use in the provided object. Default = NULL will choose the current active assay for Seurat objects and the counts assay for SingleCellExperiment objects.

use_layer

For Seurat objects, a character string or vector indicating the layer — previously known as slot — to use in the provided object. Default = NULL will use the counts layer.

n_cores

A numeric value indicating the number of cores to use for parallelization. Default = NULL will use the number of available cores minus 2.

verbose

A boolean value indicating whether to use verbose output during the execution of this function. Defaults to TRUE. Can be set to FALSE for a cleaner output.

Value

Returns a list with one pseudobulk (gene x replicate) matrix per split.

Details

This function was inspired by R package neurorestore/Libra (Squair et al. 2021).