Perform permutation test

This function performs a permutation test by shuffling the group labels for the given set of replicates and performing pseudobulk differential expression between those permuted groups. The metric used by the permutation test is the number of significantly differentially expressed features for a given significance level 'alpha'.

permuteDE(
  input,
  alpha = 0.05,
  n_iterations = 1000,
  use_splits = NULL,
  min_DE = 2,
  random_seed = 1,
  n_cores = NULL,
  verbose = TRUE
)

Arguments

input

Output from function 'runDE' containing differential expression results, pseudobulk values, and parameters used.

alpha

A numeric value indicating the significance level used for permutation test comparisons of the number of differentially expressed features. Defaults to 0.05.

n_iterations

A numeric value indicating the number of iterations run for the permutation test. Defaults to 1000.

use_splits

A vector containing the names of splits to use. Defaults to NULL.

min_DE

A numeric value indicating the minimum number of differentially expressed features between the true group labels for a split, below which permutations will not be run. Defaults to 2. Set to 0 to run permutation test for all splits, regardless of the true number of DEGs.

random_seed

A numeric value indicating the random seed to be used. Defaults to 1.

n_cores

A numeric value indicating the number of cores to use for parallelization. Default = NULL will use the number of available cores minus 2.

verbose

A boolean value indicating whether to use verbose output during the execution of this function. Defaults to TRUE. Can be set to FALSE for a cleaner output.

Value

Returns a list containing the following elements:

permutation_test_results

Dataframe containing the permutation test results by split

permutation_summary

Dataframe containing the permutation DE summary metrics by split

parameters

Dataframe record of parameter values used

Details

As input, this function requires the output from function runDE containing the pseudobulk values for each feature and the differential expression results for the true group labels.

Permutation is performed at the replicate level, rather than the cell level, and is performed without replacement, such that each iteration is a unique permutation of the group labels.

Notably, this permutation test is not performed to identify the true differentially expressed features, but rather to assess how many false positive significant differentially expressed features can be expected by chance, and to characterize the log fold change and significance observed for such false positives.