Run pseudobulk differential expression

Generate pseudobulk matri(ces) and run pseudobulk differential expression between two groups for each using existing tools: edgeR, DESeq2, or limma.

runDE(
  object,
  replicate_labels,
  group_labels,
  split_labels = NULL,
  force_balance = FALSE,
  use_cells = NULL,
  min_cells_per_split = 4,
  min_replicates_per_split = 4,
  min_replicates_per_group = 2,
  min_cells_per_feature = 1,
  de_method = "edgeR",
  de_test = "LRT",
  p_adjust_method = "fdr",
  use_assay = NULL,
  use_layer = NULL,
  n_cores = NULL,
  verbose = TRUE
)

Arguments

object

An object of class 'Seurat' or 'SingleCellExperiment'.

replicate_labels

A character string or vector indicating the name of the column containing the replicate labels.

group_labels

A character string or vector indicating the name of the column containing the two comparison group labels.

split_labels

A character string or vector indicating the name of a column by which to split the cells prior to pseudobulking and performing differential expression (e.g., cell types). Results will be returned for each unique value in the column indicated by 'split_labels'. Default = NULL will run pseudobulk differential expression on all cells together.

force_balance

A boolean indicating if two groups have equal sample size. Default to FALSE.

reference_group

A string specifying the reference group. Default to NULL, in which case the first value in the group column is used as the reference.

use_cells

A vector of cell names subset to. Default = NULL will use all cells.

min_cells_per_split

A numeric value indicating the minimum number of cells within one split. Pseudobulk expression and differential expression will not be performed for splits with fewer cells. Defaults to 4.

min_replicates_per_split

A numeric value indicating the minimum number of distinct replicates represented within one split. Pseudobulk expression and differential expression will not be performed for splits with fewer replicates. Defaults to 4.

min_replicates_per_group

A numeric value indicating the minimum number of distinct replicates represented within each of the two comparison groups. Pseudobulk expression and differential expression will not be performed for splits with fewer replicates. Defaults to 2.

min_cells_per_feature

A numeric value indicating the minimum number of cells (within a split) with expression of a gene. Pseudobulk expression and differential expression will not be calculated for genes with fewer cells. Defaults to 1.

de_method

Which tool to use for differential expression. Permitted values are 'edgeR', 'DESeq2', and 'limma'. Defaults to 'edgeR'.

de_test

Which test to use for differential expression. Defaults to 'LRT'.

p_adjust_method

A string indicating which multiple comparison adjustment to use. For permitted values, see stats::p.adjust. Defaults to 'fdr' (Benjamini & Hochberg, 1995).

use_assay

A character string indicating the assay to use in the provided object. Default = NULL will choose the current active assay for Seurat objects and the counts assay for SingleCellExperiment objects.

use_layer

For Seurat objects, a character string or vector indicating the layer — previously known as slot — to use in the provided object. Default = NULL will use the counts layer.

n_cores

A numeric value indicating the number of cores to use for parallelization. Default = NULL will use the number of available cores minus 2.

verbose

A boolean value indicating whether to use verbose output during the execution of this function. Defaults to TRUE. Can be set to FALSE for a cleaner output.

Value

Returns a list containing the following elements:

DE_results

Dataframe containing DE results for each feature, by split

PB_values

List of feature x replicate matri(ces) containing pseudobulk values for each feature, one per split

group_key

Dataframe record of group labels corresponding to each replicate

parameters

Dataframe record of parameter values used

Details

This function was inspired by R package neurorestore/Libra and some aspects are adapted therefrom (Squair et al. 2021).